app: use sentry, mv store init and provider to utils/Store
