+++ catalog/app/app.js
@@ -35,9 +35,10 @@ import { createBoundary } from 'utils/ErrorBoundary';
 import * as NamedRoutes from 'utils/NamedRoutes';
 import FormProvider from 'utils/ReduxFormProvider';
 import * as Cache from 'utils/ResourceCache';
-import StoreProvider from 'utils/StoreProvider';
+import * as Sentry from 'utils/Sentry';
+import * as Store from 'utils/Store';
 import fontLoader from 'utils/fontLoader';
-import { nest } from 'utils/reactTools';
+import { nest, composeComponent } from 'utils/reactTools';
 import RouterProvider, { LOCATION_CHANGE, selectLocation } from 'utils/router';
 import mkStorage from 'utils/storage';
 import Tracking from 'utils/tracking';
@@ -48,7 +49,6 @@ import '!file-loader?name=[name].[ext]!./manifest.json';
 import '!file-loader?name=[name].[ext]!./quilt-og.png';
 import 'file-loader?name=[name].[ext]!./.htaccess';
 /* eslint-enable import/no-unresolved, import/extensions */
-import configureStore from './store';
 // Import i18n messages
 import { translationMessages } from './i18n';
 // Import CSS reset and Global Styles
@@ -61,26 +61,28 @@ fontLoader('Roboto', 'Roboto Mono').then(() => {
   document.body.classList.add('fontLoaded');
 });
 
-// TODO: capture errors
-const ErrorBoundary = createBoundary(() => () => (
-  <Layout bare>
-    <Error
-      headline="Unexpected Error"
-      detail="Something went wrong"
-    />
-  </Layout>
-));
+const ErrorBoundary = composeComponent('ErrorBoundary',
+  Sentry.inject(),
+  createBoundary(({ sentry }) => (error, info) => {
+    sentry('captureException', error, info);
+    return (
+      <Layout bare>
+        <Error
+          headline="Unexpected Error"
+          detail="Something went wrong"
+        />
+      </Layout>
+    );
+  }));
 
-const FinalBoundary = createBoundary(() => () => (
+// error gets automatically logged to the console, so no need to do it explicitly
+const FinalBoundary = createBoundary(() => (/* error, info */) => (
   <h1 style={{ textAlign: 'center' }}>
     Something went wrong
   </h1>
 ));
 
-// Create redux store with history
-const initialState = {};
 const history = createHistory();
-const store = configureStore(initialState, history);
 const MOUNT_NODE = document.getElementById('app');
 
 // TODO: make storage injectable
@@ -95,22 +97,29 @@ const intercomUserSelector = (state) => {
   };
 };
 
+const sentryUserSelector = (state) => {
+  const { user: u } = Auth.selectors.domain(state);
+  return u ? { username: u.current_user, email: u.email } : {};
+};
+
 const render = (messages) => {
   ReactDOM.render(
     nest(
       FinalBoundary,
-      [ThemeProvider, { theme: style.theme }],
-      [StoreProvider, { store }],
+      Sentry.Provider,
+      [Store.Provider, { history }],
       [LanguageProvider, { messages }],
       [NamedRoutes.Provider, { routes }],
       [RouterProvider, { history }],
-      ErrorBoundary,
       Data.Provider,
       Cache.Provider,
       [Config.Provider, { path: '/config.json' }],
+      [ThemeProvider, { theme: style.theme }],
+      [React.Suspense, { fallback: <Placeholder /> }],
+      [Sentry.Loader, { userSelector: sentryUserSelector }],
+      ErrorBoundary,
       FormProvider,
       Notifications.Provider,
-      [React.Suspense, { fallback: <Placeholder /> }],
       [APIConnector.Provider, { fetch, middleware: [Auth.apiMiddleware] }],
       [Auth.Provider, { checkOn: LOCATION_CHANGE, storage }],
       [Intercom.Provider, { userSelector: intercomUserSelector }],
